Commits From WHSRobotics/542_16-17_fork
Due to complications with github, we were forced to create a new
repository (this one) to house our code. This is a transfer of our work
from the old repository.
I did not write all of the code in this commit.
See https://github.com/WHSRobotics/542_16-17_fork/commits/master for all
of the commits we made on the original repository.